Add & Fix 102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,7 +11,7 @@
  */
 void fibonacci(int num)
 {
-	int num1, num2, placeholder;
+	double num1, num2, placeholder;
 	num1 = 1;
 	num2 = 2;
